pyproject.toml
[project]
name = "asf-lib"
version = "0.1.0"
requires-python = ">= 3.10"
authors = [
  {name = "Hadar Shavit", email = "shavit@aim.rwth-aachen.de"},
]
maintainers = [
  {name = "Hadar Shavit", email = "shavit@aim.rwth-aachen.de"},
]
description = "Under construction! Algorithm selection framework"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research", 
  "Natural Language :: English",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  'scikit-learn',
  'pandas',
  'numpy',
]

[project.optional-dependencies]
aslib = [ 'pyyaml', 'liac-arff' ]
nn = [ 'torch' ]
xgb = [ 'xgboost' ]
tune = [ 'smac' ]
all = [ 'pyyaml', 'liac-arff', 'torch', 'xgboost', 'smac' ]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["asf*"]

[project.urls]
documentation = "https://hadarshavit.github.io/asf/latest/"
repository = "https://github.com/hadarshavit/asf"

